Single responsibility principle
