+++ config.py
@@ -1,0 +1,1 @@
+db_file = "database.db"
